Xeltalliv/clippingblending: Fix canvas alpha compatibility and other changes (#2142)
This:
* Fixes compatibility with canvas having alpha, without changing
behavior of _any_ old projects compared to how they used to work (at
least in theory).
* Sorts out long standing issue of some blending modes working
incorrectly, by deprecating old block and making a new one, where
everything is done in a corrected and clean way.
* Adds new blending modes, to provide proper solutions to what was done
through bugs.
* Adds API, so that [Shaded](https://pen-group.github.io/extensions/)
extension can be compatible with this extension, without the need for
maintaining a custom fork of it.
* Adds a label explaining why there is sometimes no blocks in the
palette.

![image](https://github.com/user-attachments/assets/774ba0c9-7373-4ad6-89ba-2afbde218f78)