Define what approving a PR means (#2327)
I don't really have the time or (more importantly) motivation to review
extensions these days.

I've gotten the feedback that reviewers don't merge pull requests
because there is no guidance when things are okay to be merged. So I'm
trying to define that here.

Here's my idea. We define passing the automated tests + getting two
approvals as fully sufficient to be merged. Now just have to define
criteria for an approval, so I've written some guidelines and a couple
templates to copy and paste into the review comment box. Check all the
boxes, then you can approve it. Get two of those, then any reviewer can
merge it.

Read the new document. It's not long. Give feedback to refine it. You'll
be the ones following it after all

---------

Co-authored-by: PPPDUD <107440101+PPPDUD@users.noreply.github.com>
Co-authored-by: DNin01 <106490990+DNin01@users.noreply.github.com>